+++ lib/src/elements/filter/convert/mod.rs
@@ -4,6 +4,10 @@
 
 #[cfg(feature = "cuda")]
 pub(crate) mod cuda;
+#[cfg(feature = "vulkan")]
+mod vulkan;
 
 #[cfg(feature = "cuda")]
 pub use cuda::{CudaConverter, CudaConverterError};
+#[cfg(feature = "vulkan")]
+pub use vulkan::{VulkanConverter, VulkanConverterError};