+++ lib/src/elements/filter/convert/vulkan/mod.rs
@@ -1,0 +1,3 @@
+mod vulkan_converter;
+
+pub use vulkan_converter::{VulkanConverter, VulkanConverterError};